requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
joblib==1.5.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
scikit-learn==1.7.1
scipy==1.16.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
